chore: move setup handler file and readline func
