+++ src/include/readline.h
@@ -1,16 +1,21 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   setup_internal.h                                   :+:      :+:    :+:   */
+/*   readline.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/01/16 08:13:10 by tookuyam          #+#    #+#             */
-/*   Updated: 2025/01/22 10:20:14 by tookuyam         ###   ########.fr       */
+/*   Created: 2025/01/22 10:09:32 by tookuyam          #+#    #+#             */
+/*   Updated: 2025/01/22 10:13:14 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef SETUP_INTERNAL_H
-# define SETUP_INTERNAL_H
+#ifndef READLINE_H
+# define READLINE_H
+
+# include "readline_var.h"
+
+char	*ms_readline(const char *prompt);
+int		ms_setup_readline_handler(void);
 
 #endif
