+++ src/setup/ms_setup.c
@@ -6,12 +6,13 @@
 /*   By: tookuyam <tookuyam@student.42tokyo.fr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/15 13:29:00 by tookuyam          #+#    #+#             */
-/*   Updated: 2025/01/19 11:19:12 by tookuyam         ###   ########.fr       */
+/*   Updated: 2025/01/22 10:18:58 by tookuyam         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "setup.h"
 #include "setup_internal.h"
+#include "readline.h"
 
 t_minishell	*ms_setup(void)
 {
@@ -19,7 +20,7 @@ t_minishell	*ms_setup(void)
 
 	ms_setup_variable();
 	ms_setup_history();
-	ms_set_handle();
+	ms_setup_readline_handler();
 	ms = ms_get_minishell();
 	*ms = (t_minishell){
 		.ms_malloc_list = NULL,
